Redirect stderr to stdout
Users can redirect the time command output to a specific file using
the following command:

        { time COMMAND; } 2>> execution_time.txt

However, if COMMAND itself also outputs to stderr, the stderr of
COMMAND will be redirected to the output file along with that of
the time command. To prevent this from happening, redirect stderr
of the command to stdout first. The modified command is as follows:

        { time COMMAND 2>&1; } 2>> execution_time.txt